+++ tests/test_unions.py
@@ -1,13 +1,13 @@
 from typing import Type, Union
 
 import attr
-from hypothesis import given
-from hypothesis.strategies import sampled_from
+import pytest
 
+from cattr._compat import is_py310_plus
 from cattr.converters import Converter, GenConverter
 
 
-@given(sampled_from([Converter, GenConverter]))
+@pytest.mark.parametrize("cls", (Converter, GenConverter))
 def test_custom_union_toplevel_roundtrip(cls: Type[Converter]):
     """
     Test custom code union handling.
@@ -40,7 +40,40 @@ def test_custom_union_toplevel_roundtrip(cls: Type[Converter]):
     assert c.structure(unstructured, Union[A, B]) == inst
 
 
-@given(sampled_from([Converter, GenConverter]))
+@pytest.mark.skipif(not is_py310_plus, reason="3.10 union syntax")
+@pytest.mark.parametrize("cls", (Converter, GenConverter))
+def test_310_custom_union_toplevel_roundtrip(cls: Type[Converter]):
+    """
+    Test custom code union handling.
+
+    We override union unstructuring to add the class type, and union structuring
+    to use the class type.
+    """
+    c = cls()
+
+    @attr.define
+    class A:
+        a: int
+
+    @attr.define
+    class B:
+        a: int
+
+    c.register_unstructure_hook(
+        A | B, lambda o: {"_type": o.__class__.__name__, **c.unstructure(o)}
+    )
+    c.register_structure_hook(
+        A | B, lambda o, t: c.structure(o, A if o["_type"] == "A" else B)
+    )
+
+    inst = B(1)
+    unstructured = c.unstructure(inst, unstructure_as=A | B)
+    assert unstructured["_type"] == "B"
+
+    assert c.structure(unstructured, A | B) == inst
+
+
+@pytest.mark.parametrize("cls", (Converter, GenConverter))
 def test_custom_union_clsfield_roundtrip(cls: Type[Converter]):
     """
     Test custom code union handling.